+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 3.7.0
 -----
 
+* remove pattern added from the Expression Violation message.
 * Added `@SerializedName` annotation support and name converters when using Symfony >= 4.2.
 
 3.3.0